refactor(core): extract decode source, probe, and channels
Move BytesMediaSource, header duration probe, and the per-channel
decode loop into audio/decode/{bytes_source,probe,channels}.rs.
Public audio::decode_* / probe_duration_* paths are unchanged.
