+++ crates/gigastt-core/src/inference/audio/decode/bytes_source.rs
@@ -1,0 +1,77 @@
+//! Zero-copy [`bytes::Bytes`] adapter for symphonia's [`MediaSource`].
+
+use bytes::Bytes;
+#[cfg(feature = "file-decode")]
+use symphonia::core::io::MediaSource;
+
+/// A [`MediaSource`] that borrows its data from a reference-counted [`Bytes`]
+/// buffer instead of cloning into a `Vec<u8>`.
+///
+/// Axum delivers REST upload bodies as `axum::body::Bytes`, which re-exports
+/// `bytes::Bytes`. Before this type the decode path called `body.to_vec()` and
+/// then wrapped the clone in `std::io::Cursor`, doubling the transient
+/// memory footprint for every upload (a 50 MiB body briefly held 100 MiB in
+/// RAM, plus another symphonia-side clone). `Bytes::clone` is a refcount bump,
+/// so the shared variant decodes the original axum buffer in place.
+///
+/// The type is deliberately small and crate-private: it only needs to satisfy
+/// `Read + Seek + Send + Sync` so symphonia's `MediaSourceStream` can drive it.
+#[allow(dead_code)] // unused when `file-decode` is off (raw-PCM-only lean build)
+pub(crate) struct BytesMediaSource {
+    data: Bytes,
+    pos: u64,
+}
+
+#[allow(dead_code)] // `new` is only called by the file-decode path
+impl BytesMediaSource {
+    pub(crate) fn new(data: Bytes) -> Self {
+        Self { data, pos: 0 }
+    }
+}
+
+impl std::io::Read for BytesMediaSource {
+    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
+        let len = self.data.len() as u64;
+        if self.pos >= len {
+            return Ok(0);
+        }
+        let start = self.pos as usize;
+        let available = self.data.len() - start;
+        let n = available.min(buf.len());
+        buf[..n].copy_from_slice(&self.data[start..start + n]);
+        self.pos += n as u64;
+        Ok(n)
+    }
+}
+
+impl std::io::Seek for BytesMediaSource {
+    fn seek(&mut self, pos: std::io::SeekFrom) -> std::io::Result<u64> {
+        let len = self.data.len() as u64;
+        // `std::io::Seek` semantics: seeking past the end is allowed; the next
+        // read returns 0. Seeking to a negative offset is an error.
+        let new_pos: i128 = match pos {
+            std::io::SeekFrom::Start(n) => n as i128,
+            std::io::SeekFrom::End(off) => len as i128 + off as i128,
+            std::io::SeekFrom::Current(off) => self.pos as i128 + off as i128,
+        };
+        if new_pos < 0 {
+            return Err(std::io::Error::new(
+                std::io::ErrorKind::InvalidInput,
+                "seek before start of buffer",
+            ));
+        }
+        self.pos = new_pos as u64;
+        Ok(self.pos)
+    }
+}
+
+#[cfg(feature = "file-decode")]
+impl MediaSource for BytesMediaSource {
+    fn is_seekable(&self) -> bool {
+        true
+    }
+
+    fn byte_len(&self) -> Option<u64> {
+        Some(self.data.len() as u64)
+    }
+}
